requirements.txt
disnake
aiohttp
random2
python-dotenv